fix: ui collections in indexer[5336]
