+++ indexer-frontend/src/app/views/collections/formulas/formulas.component.ts
@@ -20,6 +20,7 @@ import { ColumnType, TableComponent } from '@components/table/table.component';
 import { InputGroupModule } from 'primeng/inputgroup';
 import { InputGroupAddonModule } from 'primeng/inputgroupaddon';
 import { InputTextModule } from 'primeng/inputtext';
+import {HederaType} from "@components/hedera-explorer/hedera-explorer.component";
 
 @Component({
     selector: 'formulas',
@@ -53,11 +54,25 @@ import { InputTextModule } from 'primeng/inputtext';
 export class FormulasComponent extends BaseGridComponent {
     columns: any[] = [
         {
+            type: ColumnType.BUTTON,
+            title: 'grid.open',
+            btn_label: 'grid.open',
+            width: '100px',
+            callback: this.onOpen.bind(this),
+        },
+        {
             type: ColumnType.TEXT,
+            field: 'options.name',
+            title: 'grid.name',
+            width: '600px',
+        },
+        {
+            type: ColumnType.HEDERA,
             field: 'consensusTimestamp',
             title: 'grid.consensus_timestamp',
-            width: '250px',
+            width: '300px',
             sort: true,
+            hederaType: HederaType.TRANSACTION,
         },
         {
             type: ColumnType.TEXT,
@@ -71,22 +86,16 @@ export class FormulasComponent extends BaseGridComponent {
         },
         {
             type: ColumnType.TEXT,
-            field: 'options.owner',
-            title: 'grid.owner',
-            width: '650px',
-        },
-        {
-            type: ColumnType.TEXT,
-            field: 'options.name',
-            title: 'grid.name',
-            width: '200px',
-        },
-        {
-            type: ColumnType.BUTTON,
-            title: 'grid.open',
-            btn_label: 'grid.open',
-            width: '100px',
-            callback: this.onOpen.bind(this),
+            field: 'consensusTimestamp',
+            title: 'grid.date',
+            width: '250px',
+            sort: true,
+            formatValue: (value: any) => {
+                const fixedTimestamp = Math.floor(value * 1000);
+                value = new Date(fixedTimestamp);
+                const formattedDate = value.toLocaleString();
+                return formattedDate;
+            }
         },
     ];
 
@@ -97,6 +106,10 @@ export class FormulasComponent extends BaseGridComponent {
         router: Router
     ) {
         super(route, router);
+
+        this.orderField = 'consensusTimestamp';
+        this.orderDir = 'desc';
+
         this.filters.push(
             new Filter({
                 type: 'input',
